refactor(vo): vo객체 패키지 이동 및 생성자 수정
