+++ src/main/java/racingcar/domain/RacingGame.java
@@ -5,10 +5,11 @@ import java.util.List;
 import java.util.Queue;
 import java.util.stream.Collectors;
 import racingcar.InputNameHandler;
+import racingcar.domain.vo.Trial;
 import racingcar.vo.CarValue;
 import racingcar.dto.Winners;
 import racingcar.util.RandomUtil;
-import racingcar.vo.MovingNumber;
+import racingcar.domain.vo.MovingNumber;
 
 public class RacingGame {
     private static final int RANDOM_NUMBER_RANGE = 9 + 1;
@@ -18,7 +19,7 @@ public class RacingGame {
 
     public RacingGame(String carNames, int trial) {
         this.cars = new Cars(InputNameHandler.inputToNames(carNames));
-        this.trial = Trial.create(trial);
+        this.trial = new Trial(trial);
     }
 
     public void race() {
@@ -40,7 +41,7 @@ public class RacingGame {
 
     private Queue<MovingNumber> createMovingNumbers(int size) {
         return RandomUtil.getRandomNumbers(size, RANDOM_NUMBER_RANGE).stream()
-            .map(MovingNumber::create)
+            .map(MovingNumber::new)
             .collect(Collectors.toCollection(LinkedList::new));
     }
 }
