+++ src/test/java/racingcar/domain/CarTest.java
@@ -2,22 +2,19 @@ package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.vo.Name;
+import racingcar.domain.vo.Name;
 
 public class CarTest {
     @Test
     @DisplayName("이름으로 자동차 생성 테스트")
     void carTest() {
         // given
-        Name testName = Name.create("test");
+        Name testName = new Name("test");
 
         // when
         Car car = new Car(testName);
@@ -31,7 +28,7 @@ public class CarTest {
     @DisplayName("자동차 전진 테스트")
     void advanceTest(int number, int expected) {
         // given
-        Car car = new Car(Name.create("test"));
+        Car car = new Car(new Name("test"));
 
         // when
         car.advance(number);
@@ -44,7 +41,7 @@ public class CarTest {
     @ValueSource(ints = {-1, 10})
     void invalidAdvanceTest(int number) {
         // given
-        Car car = new Car(Name.create("test"));
+        Car car = new Car(new Name("test"));
 
         // when && then
         assertThatThrownBy(() -> car.advance(number))
@@ -56,8 +53,8 @@ public class CarTest {
     @DisplayName("포지션 동일성 테스트")
     void equalsPosition() {
         // given
-        Car car = new Car(Name.create("test"));
-        Car otherCar = new Car(Name.create("other"));
+        Car car = new Car(new Name("test"));
+        Car otherCar = new Car(new Name("other"));
 
         // when
         car.advance(4);
@@ -71,8 +68,8 @@ public class CarTest {
     @DisplayName("포지션이 높은 차가 양수를 반환")
     void comparePosition() {
         // given
-        Car car = new Car(Name.create("test"));
-        Car otherCar = new Car(Name.create("other"));
+        Car car = new Car(new Name("test"));
+        Car otherCar = new Car(new Name("other"));
 
         // when
         car.advance(4);
