+++ src/test/java/racingcar/domain/vo/MovingNumberTest.java
@@ -1,4 +1,4 @@
-package racingcar.vo;
+package racingcar.domain.vo;
 
 import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
@@ -12,7 +12,7 @@ class MovingNumberTest {
     @DisplayName("0과 9사이의 숫자가 아닐 경우 예외를 발생한다")
     void number_between0And9(int value) {
         assertThatIllegalArgumentException()
-            .isThrownBy(() -> MovingNumber.create(value))
+            .isThrownBy(() -> new MovingNumber(value))
             .withMessageContaining("숫자는 0과 9사이어야 합니다.");
     }
 }
