+++ src/test/java/racingcar/domain/vo/NameTest.java
@@ -1,4 +1,4 @@
-package racingcar.vo;
+package racingcar.domain.vo;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -11,7 +11,7 @@ public class NameTest {
     @CsvSource(value = {" pobi:pobi", "hoho :hoho", " rich :rich"}, delimiter = ':')
     @DisplayName("이름 앞 뒤에 공백이 존재하면 제거한다")
     void removeWhitespaces(String actual, String expected) {
-        assertThat(Name.create(actual)).isEqualTo(Name.create(expected));
+        assertThat(new Name(actual)).isEqualTo(new Name(expected));
     }
 
     @ParameterizedTest
@@ -19,7 +19,7 @@ public class NameTest {
     @DisplayName("이름이 5글자가 넘으면 예외 발생 테스트")
     void nameLengthShouldBeLessThenSix(String actual) {
         assertThatIllegalArgumentException()
-            .isThrownBy(() -> Name.create(actual))
+            .isThrownBy(() -> new Name(actual))
             .withMessageMatching("이름은 \\d+글자가 넘을 수 없습니다. :\\d+");
     }
 }
